Move code to utils
